+++ go/action_kit_commons/network/resolve_hostnames.go
@@ -10,7 +10,9 @@ import (
 	"context"
 	"fmt"
 	"net"
+	"os"
 	"slices"
+	"strconv"
 	"strings"
 
 	"github.com/rs/zerolog/log"
@@ -47,31 +49,50 @@ func (h *HostnameResolver) Resolve(ctx context.Context, hostnames ...string) ([]
 	}
 
 	if len(invalid) > 0 {
-		return nil, fmt.Errorf("could not resolve hostnames: '%s'", strings.Join(invalid, "', '"))
+		return nil, fmt.Errorf("could not resolve invalid hostnames: '%s'", strings.Join(invalid, "', '"))
 	}
 
-	outb, err := h.Dig.Run(ctx, []string{"-f-", "+timeout=4", "+noall", "+nottlid", "+answer"}, strings.NewReader(sb.String()))
-	if err != nil {
-		return nil, fmt.Errorf("could not resolve hostnames: %w", err)
-	}
+	outb, err := h.Dig.Run(ctx, []string{"-f-", fmt.Sprintf("+timeout=%d", dnsTimeout()), "+noall", "+nottlid", "+answer"}, strings.NewReader(sb.String()))
+	log.Trace().Bytes("output", outb).Err(err).Msg("dig result")
+
+	var resolved []net.IP
+	var messages []string
 
 	scanner := bufio.NewScanner(bytes.NewReader(outb))
-	var resolved []net.IP
 	for scanner.Scan() {
-		fields := strings.Fields(scanner.Text())
-		if len(fields) >= 4 {
+		line := scanner.Text()
+		if strings.HasPrefix(line, ";;") {
+			messages = append(messages, strings.TrimSpace(strings.TrimPrefix(line, ";;")))
+			continue
+		} else if fields := strings.Fields(line); len(fields) >= 4 {
 			domain := strings.TrimSuffix(fields[0], ".")
 			resolved = append(resolved, net.ParseIP(fields[3]))
 			unresolved = slices.DeleteFunc(unresolved, func(hostname string) bool {
-				return hostname == domain
+				return strings.EqualFold(hostname, domain)
 			})
 		}
+	}
+
+	if err != nil {
+		detail := ""
+		if len(messages) > 0 {
+			detail = fmt.Sprintf("\n%s", strings.Join(messages, "\n"))
+		}
+		return nil, fmt.Errorf("could not resolve hostnames: %w%s", err, detail)
 	}
 
 	if len(unresolved) > 0 {
 		return nil, fmt.Errorf("could not resolve hostnames: '%s'", strings.Join(unresolved, "', '"))
 	}
-
-	log.Trace().Interface("resolved", resolved).Strs("hostnames", hostnames).Msg("resolved resolved")
 	return resolved, nil
+}
+
+func dnsTimeout() int {
+	if s, _ := os.LookupEnv("STEADYBIT_EXTENSION_DIG_TIMEOUT"); s != "" {
+		if timeout, err := strconv.Atoi(s); err == nil && timeout > 0 {
+			return timeout
+		}
+		log.Warn().Msgf("Invalid DNS timeout value: %s, using default value of 5 seconds", s)
+	}
+	return 5
 }